// types/Messages.ts
// client/src/types/Messages.ts

export enum Message {
  UPDATE_PLAYER = 0,
  UPDATE_PLAYER_NAME = 1,
  READY_TO_CONNECT = 2,
  DISCONNECT_STREAM = 3,
  CONNECT_TO_COMPUTER = 4,
  DISCONNECT_FROM_COMPUTER = 5,
  STOP_SCREEN_SHARE = 6,
  CONNECT_TO_WHITEBOARD = 7,
  DISCONNECT_FROM_WHITEBOARD = 8,
  VIDEO_CONNECTED = 9,
  ADD_CHAT_MESSAGE = 10,
  SEND_ROOM_DATA = 11,
  CONNECT_TO_CHAIR = 12,
  DISCONNECT_FROM_CHAIR = 13,
}